% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.diet.R
\name{fb.diet}
\alias{fb.diet}
\title{Extract diet using SpecCode}
\usage{
fb.diet(SpecCode)
}
\arguments{
\item{SpecCode}{A vector of fishbase SpecCode.}
}
\value{
A matrix of diet, including:
\describe{
  \item{SpecCode}{}
  \item{no.of.diet}{}
  \item{main.food}{}
}
}
\description{
Extract diet from fishbase using SpecCode
}
\details{
This funtion extracts diet information from fishbase using SpecCode
}
\examples{
fb.diet(c(69, 309, 2420, 3272))

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}

