% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.FAO.R
\name{fb.FAO}
\alias{fb.FAO}
\title{Extract FAO using SpecCode}
\usage{
fb.FAO(SpecCode)
}
\arguments{
\item{SpecCode}{A vector of fishbase SpecCode.}
}
\value{
A matrix of FAO information, including:
\describe{
  \item{SpecCode}{}
  \item{no.of.fao}{}
  \item{fao.area}{}
}
}
\description{
Extract FAO from fishbase using SpecCode
}
\details{
This funtion extracts FAO information from fishbase using SpecCode
}
\examples{
fb.FAO(c(69, 309, 2420, 3272))

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}

