% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.fooditem.R
\name{fb.fooditem}
\alias{fb.fooditem}
\title{Extract fooditem using SpecCode}
\usage{
fb.fooditem(SpecCode)
}
\arguments{
\item{SpecCode}{A vector of fishbase SpecCode.}
}
\value{
A matrix of fooditem , including:
\describe{
  \item{SpecCode}{}
  \item{no.of.food1}{}
  \item{food1}{}
  \item{no.of.food2}{}
  \item{food2}{}
  \item{no.of.food3}{}
  \item{food3}{}
}
}
\description{
Extract fooditem from fishbase using SpecCode
}
\details{
This funtion extracts fooditem information from fishbase using SpecCode
}
\examples{
fb.fooditem(c(69, 309, 2420, 3272))

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}

