% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.mod.est.R
\name{fb.mod.est}
\alias{fb.mod.est}
\title{Extract model estimates using SpecCode}
\usage{
fb.mod.est(SpecCode)
}
\arguments{
\item{SpecCode}{A vector of fishbase SpecCode.}
}
\value{
A vector of model estimated properties, including:
\describe{
  \item{phylogenetic.diversity}{}
  \item{length.weight.a}{}
  \item{length.weight.a.min}{}
  \item{length.weight.a.max}{}
  \item{length.weight.b}{}
  \item{length.weight.b.min}{}
  \item{length.weight.b.max}{}
  \item{trophic.level}{}
  \item{trophic.level.se}{}
  \item{double.time.min}{}
  \item{double.time.max}{}
  \item{Vulnerability}{}
  \item{price}{}
}
}
\description{
Extract estimates of some fishbase properties based on models using SpecCode
}
\details{
This funtion extracts estimates of some properties based on models using SpecCode
}
\examples{
fb.mod.est(c(69, 309, 2420, 3272))

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}

